# server/docs/swagger.yaml
openapi: 3.0.0
info:
  title: TradeVault API
  description: |
    A comprehensive trading platform API that provides:

    ## Features
    - **User Authentication**: Secure login/register with JWT tokens
    - **OTP Verification**: Email-based OTP system for account security
    - **Real-time Trading**: WebSocket-based live stock updates
    - **Stock Management**: Buy/sell stocks with portfolio tracking
    - **Biometric Authentication**: Advanced security with biometric keys
    - **OAuth Integration**: Google and Apple sign-in support

    ## Authentication
    - **App Token**: Used for main application endpoints (JWT_SECRET)
    - **Socket Token**: Used for WebSocket connections (SOCKET_TOKEN_SECRET)

    ## Rate Limiting
    - OTP requests: 1 per minute per email
    - Login attempts: 3 attempts before 30-minute lockout
    - PIN attempts: 3 attempts before 30-minute lockout

  version: "1.0.0"
  contact:
    name: TradeVault Support
    email: support@tradevault.com
    url: https://tradevault.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.tradevault.com
    description: Production server
security:
  - BearerAuth: []
  - SocketAuth: []
paths:
...
  /stocks:
    get:
      tags:
        - Stock Management
      summary: Get all stocks
      description: Retrieves a list of all available stocks
      operationId: getAllStocks
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Stocks retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetStocksResponse"
        "401":
          description: Unauthorized - invalid socket token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stocks/stock:
    get:
      tags:
        - Stock Management
      summary: Get stock by symbol
      description: Retrieves detailed information about a specific stock
      operationId: getStockBySymbol
      security:
        - BearerAuth: []
      parameters:
        - name: stock
          in: query
          description: Stock symbol (e.g., AAPL, GOOGL)
          required: true
          schema:
            type: string
          example: "AAPL"
      responses:
        "200":
          description: Stock retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetStockResponse"
        "400":
          description: Bad request - missing stock symbol
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Stock not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stocks/register:
    post:
      tags:
        - Stock Management
      summary: Register new stock
      description: Adds a new stock to the trading platform
      operationId: registerStock
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterStockRequest"
            example:
              symbol: "TSLA"
              companyName: "Tesla, Inc."
              iconUrl: "https://example.com/tesla-icon.png"
              lastDayTradedPrice: "250.00"
              currentPrice: 255.50
      responses:
        "201":
          description: Stock registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterStockResponse"
        "400":
          description: Bad request - invalid data or stock already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ========================================
  # TRADING ENDPOINTS (Protected)
  # ========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for application authentication.
        Include in Authorization header: `Bearer <token>`
    SocketAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for WebSocket and trading operations.
        Include in Authorization header: `Bearer <token>`
  schemas:
...
    RegisterStockRequest:
      type: object
      required:
        - symbol
        - companyName
        - iconUrl
        - lastDayTradedPrice
        - currentPrice
      properties:
        symbol:
          type: string
          description: Stock symbol/ticker
          example: "TSLA"
        companyName:
          type: string
          description: Company name
          example: "Tesla, Inc."
        iconUrl:
          type: string
          format: uri
          description: URL to company icon
          example: "https://example.com/tesla-icon.png"
        lastDayTradedPrice:
          type: string
          description: Previous day's closing price
          example: "250.00"
        currentPrice:
          type: number
          description: Current stock price
          example: 255.50
    RegisterStockResponse:
      type: object
      properties:
        msg:
          type: string
          description: Success message
          example: "Stock added successfully!"
        data:
          $ref: "#/components/schemas/Stock"
    GetStocksResponse:
      type: object
      properties:
        msg:
          type: string
          description: Success message
          example: "Stocks retrieved successfully!"
        data:
          type: array
          items:
            $ref: "#/components/schemas/Stock"
    GetStockResponse:
      type: object
      properties:
        msg:
          type: string
          description: Success message
          example: "Stock retrieved successfully!"
        data:
          $ref: "#/components/schemas/Stock"

    # ========================================
    # TRADING SCHEMAS
    # ========================================
...
    Stock:
      type: object
      properties:
        _id:
          type: string
          description: Unique stock identifier
          example: "507f1f77bcf86cd799439011"
        symbol:
          type: string
          description: Stock symbol/ticker
          example: "TSLA"
        companyName:
          type: string
          description: Company name
          example: "Tesla, Inc."
        iconUrl:
          type: string
          format: uri
          description: URL to company icon
          example: "https://example.com/tesla-icon.png"
        lastDayTradedPrice:
          type: string
          description: Previous day's closing price
          example: "250.00"
        currentPrice:
          type: number
          description: Current stock price
          example: 255.50
        dayTimeSeries:
          type: array
          description: Intraday price data
          items:
            type: object
        tenMinTimeSeries:
          type: array
          description: 10-minute interval price data
          items:
            type: object
...
    ErrorResponse:
      type: object
      properties:
        msg:
          type: string
          description: Error message
          example: "Invalid request data"
...